updates routes path in services

+++ client/src/services/posts.js
@@ -1,25 +1,26 @@
+import api from './api-helper';
 
 export const getAllPosts = async () => {
   const resp = await api.get('/posts');
   return resp.data;
 };
 
-export const getOnePost = async (id) => {
-  const resp = await api.get(`/posts/${id}`);
+export const getOnePost = async (user_id, id) => {
+  const resp = await api.get(`/users/${user_id}/posts/${id}`);
   return resp.data;
 }
 
-export const postPost = async (postData) => {
-  const resp = await api.post('/posts', { post: postData })
+export const postPost = async (user_id, postData) => {
+  const resp = await api.post(`/users/${user_id}/posts`, { post: postData })
   return resp.data;
 }
 
-export const putPost = async (id, postData) => {
-  const resp = await api.put(`/posts/${id}`, { post: postData })
+export const putPost = async (user_id, id, postData) => {
+  const resp = await api.put(`/users/${user_id}/posts/${id}`, { post: postData })
   return resp.data;
 }
 
-export const deletePost = async (id) => {
-  const resp = await api.delete(`/posts/${id}`);
+export const deletePost = async (user_id, id) => {
+  const resp = await api.delete(`/users/${user_id}/posts/${id}`);
   return resp
 }
